Fix Asn1BitString to handle incoming data correctly
The bits in an ASN.1 BitString are NOT meant to be converted into numbers, they are flag(0) is the first/leftmost/msb in the BitString, and the numbers go UP by 1 as you go to the right. 10010 means that flag(0) and flag(4) are set. It's not meant to be interpreted as 0b10010 / 0x12 / 18dec and certainly not read in the "usual" order where those bits mean 16,8,4,2,1.